refactor: :recycle: Simplify general structure of the downloader (make it flat instead of hierarchy with one file)
