+++ boosty_downloader/src/download_manager/html_reporter.py
